Revert "Minor fixes"
This reverts commit e56175b928b952d6918e6bb2b1302ab6b4a1a6f3.

+++ src/neuralnet/layers/functionlayers/LossLayer.cpp
@@ -45,7 +45,7 @@ void LossLayer::setPlatform(Platform *platform) {
 int LossLayer::getDifficulty() {
     inputWrapper = new DataWrapper(inputDimensions);
     if (this->difficulty == 0)
-        this->difficulty = static_cast<int>(inputWrapper->getNumElements()); // Linear to input
+        this->difficulty = inputWrapper->getNumElements(); // Linear to input
     delete inputWrapper;
     return difficulty;
 }
